==> 2_enter_effect_sizes.R <==

# Notes:
#
#  ** = had raw data; see 1_get_effect_sizes_from_raw.R for how we calculated the estimates
#
#  - Analyzing on RR scale, not SMD, because for many studies, the Chinn assumption of an underlying
#   normal RV seems not reasonable. For others, the common-outcome assumption seems questionable.
#   With this modification, we assume the SMDs are computed on normal data, which seems more reasonable.
#   Used square-root transformation when outcome was rare or if we couldn't tell.
#
# - Tried to calculate RRs contrasting "high" vs. "low" meat consumption per each study's definition of those
#   terms. Sometimes these differed from study's own analyses (e.g., Anderson 2017, for which their model estimated
#   cumulative logit ORs for ordinal categories of consumption). When the outcome was about reducing vs. increasing
#   future meat consumption, we dichotomized at "staying the same".
#
# - For longitudinal controlled studies with pre/post measures, when available, used DID rather than just comparison
#  at final time point. Used latter if it was the only thing available.
#
# - For longitudinal studies with multiple post-intervention time points, we used the one with the longest
#   follow-up.
#
# - When there were multiple point estimates on same subjects, but our hierarchy didn't help decide between them,
#   we averaged the effect sizes within a study and assumed independence for a conservative SE (e.g., Hennessy).
#
# - When raw data were available, we calculated the RR for being below vs. above median meat consumption, where "median"
#   is the median at baseline among both control and treateds in a longitudinal study, or the median among controls in a
#   a 2-group study. In longitudinal studies, we controlled for meat consumption at baseline (treated as continuous).
#   When a servings variable wasn't available, we tried to instead dichotomize at "reduce meat consumption" vs. "stay
#   the same or increase" (e.g., Reese 2015).
#
#  - When different interventions in the study had different scopes (e.g., one leaflet says "go vegetarian" but another says
#   "go vegan"), defined outcome based on the most stringent intervention (e.g., vegan > vegetarian). See, e.g., Cooney 2015.


# Other notes 
#  - When possible, calculated sample size and percent male ourselves, which is why it sometimes differed from
#  what's reported in paper (e.g., due to missing data or exclusion of some conditions that weren't eligible)
#
# - Sometimes get_rr_adj's proportion of missing data differs from what we entered in
#  qualitative spreadsheet. This is intentional. get_rr_adj is only aware of missing data
#  due to observations that are in the raw data, but are dropped in analysis. Sometimes we
#  knew there was additional missing data that were not even in the raw data. 

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ #
#                                           PRELIMINARIES                                             #
# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ #

# This script uses renv to preserve the R environment specs (e.g., package versions.)
library(renv)
# run this if you want to reproduce results using the R environment we had:
# library(here); setwd(here)
# renv::restore()

library(metafor)
library(dplyr)
library(testthat)
library(readxl)
library(stringr)
library(naniar)

data.dir = "~/Dropbox/Personal computer/Independent studies/2019/AWR (animal welfare review meat consumption)/Linked to OSF (AWR)/Data extraction"
code.dir = "~/Dropbox/Personal computer/Independent studies/2019/AWR (animal welfare review meat consumption)/Linked to OSF (AWR)/Data extraction/awr_data_extraction_git"
# location of original datasets and code for reproducible studies
original.data.dir = "~/Dropbox/Personal computer/Independent studies/2019/AWR (animal welfare review meat consumption)/Literature search/Full texts for review/*INCLUDED STUDIES"
setwd(code.dir); source("helper_extraction.R")

# should we re-run the lengthy script that enters each effect size?
reenter.effect.sizes = FALSE


d = as.data.frame( matrix( ncol = 10, nrow = 0 ) )
names(d) = c( "authoryear",
              "substudy",
              "desired.direction",
              "effect.measure",
              "interpretation",
              "use.rr.analysis",
              "use.grams.analysis",
              "use.veg.analysis",
              "yi",
              "vi")

# cast data types in empty df to avoid add_row trouble
d$authoryear = as.character(d$authoryear)
d$substudy = as.character(d$substudy)
d$interpretation = as.character(d$interpretation)
d$effect.measure = as.character(d$effect.measure)

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ #
#                                  STEP 1 - MAIN-ANALYSIS STUDIES                                      #
# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ #

# asterisks denote studies with raw data

if ( reenter.effect.sizes == TRUE ) {
  
##### **Amiot 2018 (PLOS) #####
# MM audited 2020-2-2
d = dplyr::add_row(.data = d,
                   authoryear = "Amiot 2018",
                   substudy = NA,
                   desired.direction = 1,
                   effect.measure = "log-rr",
                   interpretation = "Low vs. high 1-week consumption",
                   use.rr.analysis = 1,
                   use.grams.analysis = 0,
                   use.veg.analysis = 0,
                   yi = 0.395127,
                   vi = 5.670225e-02 )


##### **Anderson 2016 (PLOS), Study 3 ######
# MM audited 2020-2-2
# same study, 3 effect sizes for different analyses
d = dplyr::add_row(.data = d,
                   authoryear = "Anderson 2016",
                   substudy = NA,
                   desired.direction = 1,
                   effect.measure = "log-rr",
                   interpretation = "Low vs. high consumption",
                   use.rr.analysis = 1,
                   use.grams.analysis = 0,
                   use.veg.analysis = 0,
                   yi = 0.1092,
                   vi = 0.0060 )

##### Schnabelrauch Arndt (2016) #####
# MM audited 2020-2-2
# total sample size in paper: 105 (per pg 52; Table 13) + 296 (per pg 95; Table 46)

# Table 13, personalized & not tailored
escalc_add_row( authoryear = "Schnabelrauch Arndt 2016",
                substudy = "Study 1, personalized & not tailored",
                desired.direction = 1,
                effect.measure = "smd",
                interpretation = "Servings SMD",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "SMD",

                m1i = 2.57,
                sd1i = 2.68,
                n1i = 37,  # sample sizes on pg 52

                m2i = 2.63,
                sd2i = 2.02,
                n2i = 40 )

# Table 13, generalized & not tailored
escalc_add_row( authoryear = "Schnabelrauch Arndt 2016",
                substudy = "Study 1, generalized & not tailored",
                desired.direction = 1,
                effect.measure = "smd",
                interpretation = "Servings SMD",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "SMD",

                m1i = 2.21,
                sd1i = 1.63,
                n1i = 29,

                m2i = 2.63,
                sd2i = 2.02,
                n2i = 40 )



# Table 46, self-schema & not tailored
# using 296/9 for all sample sizes in this table; somehow the actual breakdown isn't reported
escalc_add_row( authoryear = "Schnabelrauch Arndt 2016",
                substudy = "Study 2, self-schema & not tailored",
                desired.direction = 1,
                effect.measure = "smd",
                interpretation = "Servings SMD",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "SMD",

                m1i = 2.13,
                sd1i = 1.96,
                n1i = round(296/9),

                m2i = 2.75,
                sd2i = 2,
                n2i = round(296/9) )

# Table 46, altruistic & not tailored
escalc_add_row( authoryear = "Schnabelrauch Arndt 2016",
                substudy = "Study 2, altruistic & not tailored",
                desired.direction = 1,
                effect.measure = "smd",
                interpretation = "Servings SMD",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "SMD",

                m1i = 2.24,
                sd1i = 1.96,
                n1i = round(296/9),

                m2i = 2.75,
                sd2i = 2,
                n2i = round(296/9) )

# Table 46, egotistic & not tailored
escalc_add_row( authoryear = "Schnabelrauch Arndt 2016",
                substudy = "Study 2, egotistic & not tailored",
                desired.direction = 1,
                effect.measure = "smd",
                interpretation = "Servings SMD",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "SMD",

                m1i = 1.92,
                sd1i = 1.32,
                n1i = round(296/9),

                m2i = 2.75,
                sd2i = 2,
                n2i = round(296/9) )

# Table 46, non-specific & not tailored
escalc_add_row( authoryear = "Schnabelrauch Arndt 2016",
                substudy = "Study 2, non-specific & not tailored",
                desired.direction = 1,
                effect.measure = "smd",
                interpretation = "Servings SMD",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "SMD",

                m1i = 2.57,
                sd1i = 2.3,
                n1i = round(296/9),

                m2i = 2.75,
                sd2i = 2,
                n2i = round(296/9) )



##### **Caldwell 2016 #####

# MM audited 2020-2-2
# reduce vs. keep consumption the same or increase

# pork policy article
d = dplyr::add_row(.data = d,
                   authoryear = "Caldwell 2016",
                   substudy = "pork policy article",
                   desired.direction = 1,
                   interpretation = "Reduce pork vs. don't",
                   use.rr.analysis = 1,
                   use.grams.analysis = 0,
                   use.veg.analysis = 0,
                   effect.measure = "log-rr",
                   yi = 0.5041,
                   vi = 0.0844 )

# pork legislation
d = dplyr::add_row(.data = d,
                   authoryear = "Caldwell 2016",
                   substudy = "pork legislation article",
                   desired.direction = 1,
                   interpretation = "Reduce pork vs. don't",
                   use.rr.analysis = 1,
                   use.grams.analysis = 0,
                   use.veg.analysis = 0,
                   effect.measure = "log-rr",
                   yi = 0.7200,
                   vi = 0.0679 )


##### **Caldwell 2017a #####

# MM audited 2020-2-2

# reduce vs. keep consumption the same
# no servings variable
# for stats, used the Tableau plots on their website
#  (https://mercyforanimals.lat/kinds-of-viral-videos-are-best-at-inspiring)
#  and used the supplement for N's

# got number of control subjects from raw data (see get_effect_sizes_from_raw.R)
nc = 535

# cruelty videos
escalc_add_row( authoryear = "Caldwell 2017",
                substudy = "cruelty videos",
                desired.direction = 1,
                effect.measure = "log-rr",
                interpretation = "Reduce vs. don't",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "RR",

                ai = round(.278 * 515), # Tx intending to reduce meat
                bi = round( (1-.278) * 515),  # Tx not intending
                ci = round(.216 * nc),  # control intending to reduce
                di = round( (1-.216) * nc) )  # control not intending


# comparison videos
escalc_add_row( authoryear = "Caldwell 2017",
                substudy = "comparison videos",
                desired.direction = 1,
                effect.measure = "log-rr",
                interpretation = "Reduce vs. don't",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "RR",

                ai = round(.294 * 507), # Tx intending to reduce meat
                bi = round( (1-.294) * 507),  # Tx not intending
                ci = round(.216 * nc),  # control intending to reduce
                di = round( (1-.216) * nc) )  # control not intending


# cute videos
escalc_add_row( authoryear = "Caldwell 2017",
                substudy = "cute videos",
                desired.direction = 1,
                effect.measure = "log-rr",
                interpretation = "Reduce vs. don't",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "RR",

                ai = round(.24 * 513), # Tx intending to reduce meat
                bi = round( (1-.24) * 513),  # Tx not intending
                ci = round(.216 * nc),  # control intending to reduce
                di = round( (1-0.216) * nc) )  # control not intending

# lifestyle videos aren't animal welfare



##### Cordts 2014 ######
# MM audited 2020-2-2
# Table 6
escalc_add_row( authoryear = "Cordts 2014",
                substudy = NA,
                desired.direction = 1,
                effect.measure = "log-rr",
                interpretation = "Reduce vs. don't",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "RR",

                ai = round(.28 * 150), # Tx intending to reduce meat
                bi = round( (1-.28) * 150),  # Tx not intending
                ci = round(.128 * 556),  # control intending to reduce
                di = round( (1-.128) * 556) )  # control not intending
# sanity check
expect_equal( exp( d$yi[ d$authoryear == "Cordts 2014" ] ),
              (42/150) / (71/556) )
# mean age from table with categories
(.053*mean(c(18,24)) + .261*mean(c(25,39)) + .523*mean(c(40,59)) + .163*60)/(.053+.261+.523+.163)


##### **Palomo-Velez 2018 #####
# MM audited 2020-2-2
# total sample size for paper: 151 + 161 + 270 from raw data
# public data; used median split

# Study 1
d = dplyr::add_row(.data = d,
                   authoryear = "Palomo-Velez 2018",
                   substudy = "Study 1",
                   desired.direction = 1,
                   interpretation = "Low vs. high meat buying likelihood",
                   use.rr.analysis = 1,
                   use.grams.analysis = 0,
                   use.veg.analysis = 0,
                   effect.measure = "log-rr",
                   yi = 0.2319,
                   vi = 0.0230 )


# Study 2
d = dplyr::add_row(.data = d,
                   authoryear = "Palomo-Velez 2018",
                   substudy = "Study 2",
                   desired.direction = 1,
                   interpretation = "Low vs. high meat buying likelihood",
                   use.rr.analysis = 1,
                   use.grams.analysis = 0,
                   use.veg.analysis = 0,
                   effect.measure = "log-rr",
                   yi = 0.1133,
                   vi = 0.0259 )


# Study 3
d = dplyr::add_row(.data = d,
                   authoryear = "Palomo-Velez 2018",
                   substudy = "Study 3",
                   desired.direction = 1,
                   interpretation = "Low vs. high meat buying likelihood",
                   use.rr.analysis = 1,
                   use.grams.analysis = 0,
                   use.veg.analysis = 0,
                   effect.measure = "log-rr",
                   yi = 0.2020,
                   vi = 0.0123 )


##### **Anderson 2017 (iAnimal), 360 video and 2D video ######

# MM audited 2020-2-2
d = dplyr::add_row(.data = d,
                   authoryear = "Anderson 2017",
                   substudy = "standard video",
                   desired.direction = 1,
                   effect.measure = "log-rr",
                   interpretation = "Low vs. high consumption frequency",
                   use.rr.analysis = 1,
                   use.grams.analysis = 0,
                   use.veg.analysis = 0,
                   yi = 0.17653,
                   vi = 0.006292858 )

d = dplyr::add_row(.data = d,
                   authoryear = "Anderson 2017",
                   substudy = "360-degree video",
                   desired.direction = 1,
                   effect.measure = "log-rr",
                   interpretation = "Low vs. high consumption frequency",
                   use.rr.analysis = 1,
                   use.grams.analysis = 0,
                   use.veg.analysis = 0,
                   yi = 0.07114,
                   vi = 0.001103623 )


##### Bertolaso 2015 #####
# MM audited 2020-2-2

# pre-post design, but they only report marginal means at each time, not change scores,
# so we are only using post-intervention means
# all stats in Table 1 (pg. 53)

# moral shocks & promotion 
escalc_add_row( authoryear = "Bertolaso 2015",
                substudy = "moral shocks & promotion focus",
                desired.direction = 1,
                effect.measure = "smd",
                interpretation = "Consumption frequency SMD",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "SMD",

                m1i = 1.6, # Table 1, control group
                sd1i = 0.38,
                n1i = 107,

                m2i = 1.48,
                sd2i = 0.38,
                n2i = 106 )

# moral shocks & prevention
escalc_add_row( authoryear = "Bertolaso 2015",
                substudy = "moral shocks & prevention focus",
                desired.direction = 1,
                effect.measure = "smd",
                interpretation = "Consumption frequency SMD",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "SMD",

                m1i = 1.6, # Table 1
                sd1i = 0.38,
                n1i = 107,

                m2i = 1.51,
                sd2i = 0.33,
                n2i = 108 )


# individualization & promotion
escalc_add_row( authoryear = "Bertolaso 2015",
                substudy = "individualization & promotion focus",
                desired.direction = 1,
                effect.measure = "smd",
                interpretation = "Consumption frequency SMD",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "SMD",

                m1i = 1.6, # Table 1
                sd1i = 0.38,
                n1i = 107,

                m2i = 1.48,
                sd2i = 0.31,
                n2i = 92 )

# individualization & prevention
escalc_add_row( authoryear = "Bertolaso 2015",
                substudy = "individualization & prevention focus",
                desired.direction = 1,
                effect.measure = "smd",
                interpretation = "Consumption frequency SMD",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "SMD",

                m1i = 1.6, # Table 1
                sd1i = 0.38,
                n1i = 107,

                m2i = 1.55,
                sd2i = 0.35,
                n2i = 98 )


##### **Cooney 2014 #####
# MM audited 2020-2-2
# this one had a ton of effect sizes, so was prepped more automatically
setwd(original.data.dir)
setwd("Cooney 2014, #3856")

d = rbind( d, read.csv("cooney_2014_prepped_effect_sizes.csv")[,-1] )


##### **Doebel 2015 #####
# MM audited 2020-2-2
# this one had a ton of effect sizes, so was prepped more automatically
setwd(original.data.dir)
setwd("Doebel 2015, #3866")

d = rbind( d, read.csv("doebel_2015_prepped_effect_sizes.csv")[,-1] )


##### **Cooney 2016 #####
# MM audited 2020-2-2
d = dplyr::add_row(.data = d,
                   authoryear = "Cooney 2016",
                   substudy = NA,
                   desired.direction = 0,
                   effect.measure = "log-rr",
                   interpretation = "Low vs. high 2-day consumption",
                   use.rr.analysis = 1,
                   use.grams.analysis = 0,
                   use.veg.analysis = 0,
                   yi = -0.0107,
                   vi = 0.0026 )


##### Hennessy 2016, "why" leaflet #####
# MM audited 2020-2-2
# stats in Table 3
# "Total" = sum of number of days over past week on which subject ate various animal products

# control N: 170 (page 30)
# "why" leaflet (#1) N: 79 (page 23)
# "how" leaflet (#2) N: 78 (page 23)

# number of days subject ate meat over past week + similar for dairy + similar for eggs
escalc_add_row( authoryear = "Hennessy 2016",
                substudy = 'why leaflet',
                desired.direction = 0,
                effect.measure = "smd",
                interpretation = "Animal product consumption frequency SMD",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "SMD",

                m1i = 12.22, # Table 3
                sd1i = 4.25,
                n1i = 79,

                m2i = 11.64,  # control group
                sd2i = 4.14,
                n2i = 170 )


escalc_add_row( authoryear = "Hennessy 2016",
                substudy = 'how leaflet',
                desired.direction = 0,
                effect.measure = "smd",
                interpretation = "Animal product consumption frequency SMD",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "SMD",

                m1i = 12.65, # Table 3
                sd1i = 3.92,
                n1i = 78,

                m2i = 11.64,  # control group
                sd2i = 4.14,
                n2i = 170 )


##### **MacDonald 2016 #####

# MM audited 2020-2-2
d = dplyr::add_row(.data = d,
                   authoryear = "MacDonald 2016",
                   substudy = "reduce",
                   desired.direction = 1,
                   effect.measure = "log-rr",
                   interpretation = "Low vs. high 1-week consumption",
                   use.rr.analysis = 1,
                   use.grams.analysis = 0,
                   use.veg.analysis = 0,
                   yi = 0.060872,
                   vi = 0.002475808 )

d = dplyr::add_row(.data = d,
                   authoryear = "MacDonald 2016",
                   substudy = "eliminate",
                   desired.direction = 1,
                   effect.measure = "log-rr",
                   interpretation = "Low vs. high 1-week consumption",
                   use.rr.analysis = 1,
                   use.grams.analysis = 0,
                   use.veg.analysis = 0,
                   yi = 0.077449,
                   vi = 0.002377879 )



##### **Reese 2015 #####
# MM audited 2020-2-3

# this one had a ton of effect sizes, so was prepped more automatically
setwd(original.data.dir)
setwd("Reese 2015, #3787")

d = rbind( d, read.csv("reese_prepped_effect_sizes.csv")[,-1] )


##### **Kunst 2016 #####
# MM audited 2020-2-3

# total N from raw data: 168 + 101 + 187 + 190 = 646
d = dplyr::add_row(.data = d,
                   authoryear = "Kunst 2016",
                   substudy = "Study 2A",
                   desired.direction = 1,
                   effect.measure = "log-rr",
                   interpretation = "Low vs. high willingness to eat specific meat sample",
                   use.rr.analysis = 1,
                   use.grams.analysis = 0,
                   use.veg.analysis = 0,
                   yi = 0.3735,
                   vi = 0.0170 )

d = dplyr::add_row(.data = d,
                   authoryear = "Kunst 2016",
                   substudy = "Study 2B",
                   desired.direction = 1,
                   effect.measure = "log-rr",
                   interpretation = "Low vs. high willingness to choose vegetarian dish",
                   use.rr.analysis = 1,
                   use.grams.analysis = 0,
                   use.veg.analysis = 0,
                   yi = 0.3167,
                   vi = 0.0290 )

d = dplyr::add_row(.data = d,
                   authoryear = "Kunst 2016",
                   substudy = "Study 3",
                   desired.direction = 1,
                   effect.measure = "log-rr",
                   interpretation = "Low vs. high willingness to eat specific meat sample",
                   use.rr.analysis = 1,
                   use.grams.analysis = 0,
                   use.veg.analysis = 0,
                   yi = 0.3866,
                   vi = 0.0148 )

d = dplyr::add_row(.data = d,
                   authoryear = "Kunst 2016",
                   substudy = "Study 5",
                   desired.direction = 1,
                   effect.measure = "log-rr",
                   interpretation = "Low vs. high willingness to choose vegetarian dish",
                   use.rr.analysis = 1,
                   use.grams.analysis = 0,
                   use.veg.analysis = 0,
                   yi = 0.1361,
                   vi = 0.0184 )


##### **Kunst 2018 #####
# MM audited 2020-2-3

# total N from raw data: 178 + 183 = 361
# American sample
d = dplyr::add_row(.data = d,
                   authoryear = "Kunst 2018",
                   substudy = "American sample",
                   desired.direction = 1,
                   effect.measure = "log-rr",
                   interpretation = "Low vs. high willingness to choose vegetarian dish",
                   use.rr.analysis = 1,
                   use.grams.analysis = 0,
                   use.veg.analysis = 0,
                   yi = 0.3909,
                   vi = 0.0155 )

# Ecuadorian sample
d = dplyr::add_row(.data = d,
                   authoryear = "Kunst 2018",
                   substudy = "Ecuadorian sample",
                   desired.direction = 1,
                   effect.measure = "log-rr",
                   interpretation = "Low vs. high willingness to choose vegetarian dish",
                   use.rr.analysis = 1,
                   use.grams.analysis = 0,
                   use.veg.analysis = 0,
                   yi = 0.1714,
                   vi = 0.0189 )



##### Silva (2016) #####
# MM audited 2020-2-3

# measured Fig 1.2 in pixels (obviously arbitrary since it depends on zooming,
#  but it's the relative bar measurements that matter)
# bar heights in pixels:
# high cuteness: 217
# low cuteness: 209
# control: 180
# distance between "1" and "5" tick marks on y-axis: 481 px

# intervention does not make any request, so using the "avoid meat" outcome per our hierarchy

# start at 1, the lowest point on x-axis, and add the proportion of the
#  x-axis occupied by the bar
( mean.hi.cute = 1 + (217 / 481) * (5-1) )
( mean.lo.cute = 1 + (209 / 481) * (5-1) )
( mean.cntrl = 1 + (180 / 481) * (5-1) )
# conservatively use marginal SD for all groups (pg. 22) since it's all we have
sd.marg = 1.33 

escalc_add_row( authoryear = "Silva 2016",
                substudy = "cute pig",
                desired.direction = 1,
                effect.measure = "smd",
                interpretation = "Willingness to avoid meat",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "SMD",

                m1i = mean.hi.cute, # Table 1
                sd1i = sd.marg,
                n1i = 68,  # all n's on pg 15

                m2i = mean.cntrl,
                sd2i = sd.marg,
                n2i = 62 )

escalc_add_row( authoryear = "Silva 2016",
                substudy = "non-cute pig",
                desired.direction = 1,
                effect.measure = "smd",
                interpretation = "Willingness to avoid meat",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "SMD",

                m1i = mean.lo.cute, # Table 1
                sd1i = sd.marg,
                n1i = 71,  # all n's on pg 15

                m2i = mean.cntrl,
                sd2i = sd.marg,
                n2i = 62 )


##### Spanikova (2015) #####
# MM audited 2020-2-3

# n on pg. 15: 178 for restaurant task
# assume equal sample size in each of 8 groups described in Table 1
n.group = floor(178/8)  

# Table 5 (a weird, not-saturated logistic regression model)
expit = function(x) { 1 / ( 1 + exp(-x) ) }

# P( choose vegetarian | control )
( p.cntrl = expit( -1.12 + 0.95 ) )
# P( choose vegetarian | positive animal welfare intervention )
( p.pos = expit( -1.12 + 0.61 ) )
# P( choose vegetarian | negative animal welfare intervention )
( p.neg = expit( -1.12 + 0.61 - 0.33 ) )

escalc_add_row( authoryear = "Spanikova 2015",
                substudy = "positive framing",
                desired.direction = 0,
                effect.measure = "log-rr",
                interpretation = "Chose vegetarian vs. non-vegetarian restaurant",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "RR",

                ai = round(p.pos * n.group), # Tx choosing vegetarian restaurant
                bi = round( (1-p.pos) * n.group),  # Tx not choosing veg
                ci = round(p.cntrl * n.group),  # control choosing veg restaurant
                di = round( (1-p.cntrl) * n.group) )  # control not choosing veg

escalc_add_row( authoryear = "Spanikova 2015",
                substudy = "negative framing",
                desired.direction = 0,
                effect.measure = "log-rr",
                interpretation = "Chose vegetarian vs. non-vegetarian restaurant",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "RR",

                ai = round(p.neg * n.group), # Tx choosing vegetarian restaurant
                bi = round( (1-p.neg) * n.group),  # Tx not choosing veg
                ci = round(p.cntrl * n.group),  # control choosing veg restaurant
                di = round( (1-p.cntrl) * n.group) )  # control not choosing veg


##### **Norris 2016 #####

# data were prepped in get_effect_sizes_from_raw.R
setwd(original.data.dir)
setwd('Norris 2016, #3829')
setwd("Data from author/Pay Per Read Fall 2015")

d = rbind( d, read.csv("norris_2016_prepped_effect_sizes.csv")[,-1] )


##### **Norris 2019 #####
# MM audited 2020-2-3

# Compassionate Choices
d = dplyr::add_row(.data = d,
                   authoryear = "Norris 2019",
                   substudy = '"Compassionate Choices"',
                   desired.direction = 1,
                   effect.measure = "log-rr",
                   interpretation = "Low vs. high animal product consumption",
                   use.rr.analysis = 1,
                   use.grams.analysis = 0,
                   use.veg.analysis = 0,
                   yi = 0.05990,
                   vi = 6.15334e-03 )

# Speciesism
d = dplyr::add_row(.data = d,
                   authoryear = "Norris 2019",
                   substudy = '"Speciesism"',
                   desired.direction = 1,
                   effect.measure = "log-rr",
                   interpretation = "Low vs. high animal product consumption",
                   use.rr.analysis = 1,
                   use.grams.analysis = 0,
                   use.veg.analysis = 0,
                   yi = 0.16431,
                   vi = 0.0060461344 )


##### Byrd-Bredbenner 2010 #####
# MM audited 2020-2-3

# they measured pre-post changes, but change scores aren't reported, so
#  just using means at F/U
escalc_add_row( authoryear = "Byrd-Bredbenner 2010",
                substudy = NA,
                desired.direction = 1,
                effect.measure = "smd",
                interpretation = "Intention to go vegetarian",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "SMD",

                m1i = 1.94, # Table 1, treatment group
                sd1i = 0.13 * sqrt(34),  # convert SE to SD
                n1i = 34,  # pg 45 (beginning of Results)

                m2i = 1.66,
                sd2i = 0.09 * sqrt(37),
                n2i = 37 ) # pg 45 (beginning of Results)


##### Novotna 2019 #####
# MM audited 2020-2-3

# mean age from categorical breakdown on page 25
ns = c(80, 17, 1)
midpoints = c( mean(c(18,25)), mean(c(26,40)), mean(c(41,59)) )
sum(ns * midpoints) / sum(ns)

# extract post-intervention means and CI limits
#  in WebPlotDigitizer
mn.cntrl = 2.96
hw.cntrl = (mn.cntrl-2.77)  # CI half-width
n.cntrl = 34-8  # pg 26
sd.cntrl = ( hw.cntrl / qt(.975, df = n.cntrl-1) ) * sqrt(n.cntrl)
# sanity check
#qt(.975, df = n.cntrl-1) * (sd.cntrl / sqrt(n.cntrl)); hw.cntrl

mn.doc = 2.71
hw.doc = (mn.doc-2.54)
n.doc = 33-3  # pg 26
sd.doc = ( hw.doc / qt(.975, df = n.doc-1) ) * sqrt(n.doc)

mn.babe = 2.88
hw.babe = (mn.babe-2.71)
n.babe = 32-1  # pg 26
sd.babe = ( hw.babe / qt(.975, df = n.babe-1) ) * sqrt(n.babe)


escalc_add_row( authoryear = "Novotna 2019",
                substudy = "Documentary",
                desired.direction = 1,
                effect.measure = "smd",
                interpretation = "Meat and dairy consumption",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "SMD",

                m1i = mn.doc, # Figure 12
                sd1i = sd.doc,
                n1i = n.doc,

                m2i = mn.cntrl,
                sd2i = sd.cntrl,
                n2i = n.cntrl )

escalc_add_row( authoryear = "Novotna 2019",
                substudy = '"Babe" movie',
                desired.direction = 1,
                effect.measure = "smd",
                interpretation = "Meat and dairy consumption",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "SMD",

                m1i = mn.babe, # Figure 13
                sd1i = sd.babe,
                n1i = n.babe,

                m2i = mn.cntrl,
                sd2i = sd.cntrl,
                n2i = n.cntrl)


##### Tian (2016) #####
# MM audited 2020-2-3

# total sample size for paper
# divide by 2 because we're only using 2 of 4 conditions for each study (not pasture or abbatoir conditions)
round( 277/2 + 243/2 +  # Study 1 (page 189)
         301/2 + 217/2 )  # Study 2 (page 191)

# Study 1, Chinese (Table 1, abbatoir condition)
escalc_add_row( authoryear = "Tian 2016",
                substudy = "Study 1, Chinese",
                desired.direction = 1,
                effect.measure = "smd",
                interpretation = "Willingness to eat beef",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "SMD",

                m1i = 3.28,
                sd1i = 1.46,
                n1i = round(277/4),  # 4 total conditions, though we are excluding 3 of them

                m2i = 3.68,
                sd2i = 1.28,
                n2i = round(277/4) )


# Study 1, French (Table 1, abbatoir condition)
escalc_add_row( authoryear = "Tian 2016",
                substudy = "Study 1, French",
                desired.direction = 1,
                effect.measure = "smd",
                interpretation = "Willingness to eat beef SMD",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "SMD",

                m1i = 3.55,
                sd1i = 1.30,
                n1i = round(243/4),

                m2i = 3.78,
                sd2i = 1.37,
                n2i = round(243/4) )

# Study 2, Chinese (Table 3, animal image condition)
escalc_add_row( authoryear = "Tian 2016",
                substudy = "Study 2, Chinese",
                desired.direction = 0,
                effect.measure = "smd",
                interpretation = "Willingness to eat beef",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "SMD",

                m1i = 3.61,
                sd1i = 1.29,
                n1i = round(217/4),  # again, 4 total conditions, but we are only using one of the treatments

                m2i = 3.53,
                sd2i = 1.36,
                n2i = round(217/4) )

# Study 2, French (Table 3, animal image condition)
escalc_add_row( authoryear = "Tian 2016",
                substudy = "Study 2, French",
                desired.direction = 0,
                effect.measure = "smd",
                interpretation = "Willingness to eat beef",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "SMD",

                m1i = 3.84,
                sd1i = 1.30,
                n1i = round(301/4),   # again, 4 total conditions, but we are only using one of the treatments

                m2i = 3.82,
                sd2i = 1.24,
                n2i = round(301/4) )


##### Earle 2019 #####
# MM audited 2020-2-3

# emailed to try to get means and SDs, 
#  but stuck with converting the point-biserial correlation

# Study 1 (Table 1)
res = r_to_d_ptbis( r = -0.31,
              n0 = round(299/2), n1 = round(299/2) )
d = dplyr::add_row(.data = d,
                   authoryear = "Earle 2019",
                   substudy = "Study 1",
                   desired.direction = 1,
                   effect.measure = "smd",
                   interpretation = "Willingness to eat meat",
                   use.rr.analysis = 1,
                   use.grams.analysis = 0,
                   use.veg.analysis = 0,
                   yi = res$d,
                   vi = res$se^2 )

# Study 2 (Table 3)
res = r_to_d_ptbis( r = -0.35,
                    n0 = round(280/2), n1 = round(280/2) )
d = dplyr::add_row(.data = d,
                   authoryear = "Earle 2019",
                   substudy = "Study 2",
                   desired.direction = 1,
                   effect.measure = "smd",
                   interpretation = "Willingness to eat meat",
                   use.rr.analysis = 1,
                   use.grams.analysis = 0,
                   use.veg.analysis = 0,
                   yi = res$d,
                   vi = res$se^2 )


##### Flens (2018) #####
# MM audited 2020-2-4

# Table 1 (page 4)
# create aggregated measure: consumption of all animal products
#  by summing each category they report
# per page 2, negative values represent decreases in consumption; positive indicate increases
( sum.trt = sum( c( 0.34, -0.04, -0.90, 1.21, 0.57, 0.12 ) ) )
( sum.cntrl = sum( c( 0.42, -0.01, -0.88, 0.84, 0.52, 0.20 ) ) )
# assume independence for conservative pooled SE
# emailed to try to get non-conservative SD, but to no avail
( sd.trt.pooled = sqrt( 1.57^2 + 1.58^2 + 1.24^2 + 1.27^2 + 1.62^2 + 1.35^2 ) )
( sd.cntrl.pooled = sqrt( 1.40^2 + 1.27^2 + 1.55^2 + 1.30^2 + 1.43^2 + 1.41^2 ) )

escalc_add_row( authoryear = "Flens 2018",
                substudy = NA,
                desired.direction = 0,
                effect.measure = "smd",
                interpretation = "Animal products consumption frequency",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "SMD",

                m1i = sum.trt,
                sd1i = sd.trt.pooled,
                n1i = 52,  # page 3

                m2i = sum.cntrl,
                sd2i = sd.cntrl.pooled,
                n2i = 200 )  # page 3

##### **Schwitzgebel (2019) #####
# MM audited 2020-2-4
d = dplyr::add_row(.data = d,
                   authoryear = "Schwitzgebel 2019",
                   desired.direction = 1,
                   effect.measure = "log-rr",
                   interpretation = "Purchase contained meat",
                   use.rr.analysis = 1,
                   use.grams.analysis = 0,
                   use.veg.analysis = 0,
                   yi = -0.13597,
                   vi = 0.007144952 )


##### **FIAPO (2017) #####
# MM audited 2020-2-4

(91-2) + (188-21) + # Study 1, baseline non-vegans in control group (excluded the baseline vegans)
1000 + # from raw data
1040  # from raw data

# Study 1: leafleting
# combine data in Table A and Table 1 to get RR of *becoming* vegan
#  among initial non-vegans
escalc_add_row( authoryear = "FIAPO 2017",
                substudy = "leaflet",
                desired.direction = 1,
                effect.measure = "log-rr",
                interpretation = "Going vegan",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "RR",

                # proportions from Table A; N's from Table 1
                ai = round( 0.005*21 + 0.0266*146 ), # Tx who *went* vegan, of initial non-vegans
                bi = round( (188-21) - round(0.005*21 + 0.0266*146) ),  # Tx who did not *go* vegan, of initial non-vegans
                ci = 0,  # control who went vegan
                di = 91-2 ) # control who did not go vegan
                
# Study 2: 2D video
# from raw data
escalc_add_row( authoryear = "FIAPO 2017",
                substudy = "2D video",
                desired.direction = 1,
                effect.measure = "smd",
                interpretation = "Willingness to reduce meat",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "SMD",

                m1i = 4.847619,
                sd1i = 1.89707,
                n1i = 525,

                m2i = 3.95534,
                sd2i = 2.048048,
                n2i = 515 )

# Study 3: virtual reality
# from raw data
escalc_add_row( authoryear = "FIAPO 2017",
                substudy = "virtual reality",
                desired.direction = 1,
                effect.measure = "smd",
                interpretation = "Willingness to reduce meat",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "SMD",

                m1i = 3.707052,
                sd1i = 2.594777,
                n1i = 553,

                m2i = 6.550336,
                sd2i = 2.983744,
                n2i = 447 )


##### Rouk 2017 #####
# MM audited 2020-2-4

# this one had a ton of effect sizes, so was prepped more automatically
setwd(original.data.dir)
setwd("Rouk 2017, #3832")

d = rbind( d, read.csv("rouk_prepped_effect_sizes.csv")[,-1] )



##### **ACE 2013a #####

# "Compassionate Choices"
d = dplyr::add_row(.data = d,
                   authoryear = "ACE 2013a",
                   substudy = '"Compassionate Choices"',
                   desired.direction = 0,
                   effect.measure = "log-rr",
                   interpretation = "Low vs. high animal product consumption",
                   use.rr.analysis = 1,
                   use.grams.analysis = 0,
                   use.veg.analysis = 0,
                   yi = -0.3883008,
                   vi = 0.1138651 )

# "Even if You Like Meat"
d = dplyr::add_row(.data = d,
                   authoryear = "ACE 2013a",
                   substudy = '"Even If You Like Meat"',
                   desired.direction = 1,
                   effect.measure = "log-rr",
                   interpretation = "Low vs. high animal product consumption",
                   use.rr.analysis = 1,
                   use.grams.analysis = 0,
                   use.veg.analysis = 0,
                   yi = 0.001426734,
                   vi = 0.008177298 )


##### **ACE 2013b #####
# MM audited 2020-2-4
d = dplyr::add_row(.data = d,
                   authoryear = "ACE 2013b",
                   desired.direction = 0,
                   effect.measure = "smd",
                   interpretation = "Low vs. high animal product consumption",
                   use.rr.analysis = 1,
                   use.grams.analysis = 0,
                   use.veg.analysis = 0,
                   yi = -0.08581416,
                   vi = 0.01900103 )


##### Feltz 2019, #3858 #####
# MM audited 2020-2-5

# stats and sample sizes on page 11
# this is the SMD of the gain scores themselves
d = escalc_add_row( authoryear = "Feltz 2019",
                    substudy = NA,
                    desired.direction = 0,
                    effect.measure = "smd",
                    interpretation = "Within-subject change in animal product consumption",
                    use.rr.analysis = 1,
                    use.grams.analysis = 0,
                    use.veg.analysis = 0,
                    
                    measure = "SMD",
                    m1i = 0.14,  # intervention 
                    sd1i = 1.82,
                    n1i = 58, 
                    
                    m2i = -0.04,  # control
                    sd2i = 1.44,
                    n2i = 43 )


##### **Lackner 2019, #3858 #####

# MM audited 2020-2-5
d = dplyr::add_row(.data = d,
                   authoryear = "Lackner 2019",
                   desired.direction = 0,
                   effect.measure = "log-rr",
                   interpretation = "Low vs. high intended meat consumption",
                   use.rr.analysis = 1,
                   use.grams.analysis = 0,
                   use.veg.analysis = 0,
                   yi = -0.3542,
                   vi = 0.0817 )


###### **Norris 2014 #####
# MM audited 2020-2-5

# "Your Choice"
d = dplyr::add_row(.data = d,
                   authoryear = "Norris 2014",
                   substudy = '"Your Choice"',
                   desired.direction = 0,
                   effect.measure = "log-rr",
                   interpretation = "Low vs. high animal product consumption",
                   use.rr.analysis = 1,
                   use.grams.analysis = 0,
                   use.veg.analysis = 0,
                   yi = -0.108307,
                   vi = 0.01944182 )

# "Even If You Like Meat"
d = dplyr::add_row(.data = d,
                   authoryear = "Norris 2014",
                   substudy = '"Even If You Like Meat"',
                   desired.direction = 0,
                   effect.measure = "log-rr",
                   interpretation = "Low vs. high animal product consumption",
                   use.rr.analysis = 1,
                   use.grams.analysis = 0,
                   use.veg.analysis = 0,
                   yi = -0.05842669,
                   vi = 0.02098244 )


# ##### deLanauze 2019, #4033 #####
# 
# # get SDs from p-values and means
# 
# p = 
# 
# sd.cntrl = ( hw.cntrl / qt(.975, df = n.cntrl-1) ) * sqrt(n.cntrl)
# # sanity check
# 
# d = dplyr::add_row(.data = d,
#                    authoryear = "deLanauze 2019",
#                    desired.direction = ,
#                    effect.measure = "smd",
#                    interpretation = "",
#                    use.rr.analysis = 1,
#                    use.grams.analysis = 0,
#                    use.veg.analysis = 0,
#                    yi = ,
#                    vi = )
# 
# 
# # confirm equivalence of transformation
# n0 = 100
# n1 = 100
# y0 = rnorm(n=n0, mean=0, sd=2)
# y1 = rnorm(n=n1, mean=y0 + 1, sd=2)  # make them correlated
# diff = y1-y0
# 
# # independent-samples t-test
# t1 = t.test(y0, y1)$statistic
# 
# # paired t-test
# t2 = t.test(y0, y1, paired = TRUE)$statistic
# 
# 
# # directly calculate d
# escalc( measure = "SMD",
#                 
#                 m1i = mean(y0),
#                 sd1i = sd(y0),
#                 n1i = n0,  # sample sizes on pg 52
#                 
#                 m2i = mean(y1),
#                 sd2i = sd(y1),
#                 n2i = n1 )
# 
# t * sqrt( (n0+n1) / (n0*n1) )
# 
# # they did a paired t-test, but we're treating as independent, so t-value we use
#  will be larger than it should be, so resulting d will be larger than what we 
#  would have gotten with direct independent calculation


# save intermediate dataset
setwd(data.dir)
write.csv(d, "data_prepped_step1.csv", row.names = FALSE)

} # end massive loop "if (reenter.effect.sizes == TRUE)"




# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ #
#                                     STEP 2 - SSWS                                      #
# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ #

if ( reenter.effect.sizes == TRUE ) {
  
# read it back in
setwd(data.dir)
d = read.csv("data_prepped_step1.csv")


##### #3865 30-day VeggieChallenge 2016 (ProVeg) #####

# Mensink Table 2a
escalc_add_row( authoryear = "30-Day VeggieChallenge (ProVeg) 2016",
                substudy = NA,
                desired.direction = 1,
                effect.measure = "log-rr",
                interpretation = "Being vegan after vs. before intervention",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "RR",

                ai = 14, # pre-intervention vegans
                bi = 60+20+102+32+60+15+17,  # pre-intervention non-vegans (2016)
                ci = 36,  # post- vegans
                di = 17+16+72+54+57+48+21 ) # post-non-vegans (2016)
14/(14+60+20+102+32+60+15+17)
36/(17+16+72+54+57+48+21)


##### #3865 30-day VeggieChallenge 2015 (ProVeg) #####
# MM audited 2020-2-5

# Mensink Table 2b
escalc_add_row( authoryear = "30-Day VeggieChallenge (ProVeg) 2015",
                substudy = NA,
                desired.direction = 1,
                effect.measure = "log-rr",
                interpretation = "Being vegan after vs. before intervention",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "RR",

                ai = 15, # pre-intervention vegans
                bi = 21+40+27+48,  # pre-intervention non-vegans
                ci = 17,  # post- vegans
                di = 4+29+32+35 ) # post-non-vegans

# sanity check
15/(15+21+40+27+48)
17/(17+4+29+32+35)


##### #3827 30-day VeggieChallenge 2017 (ProVeg) #####
# MM audited 2020-2-5

# Moleman 2018's Table 6:
# pre-intervention: 4% vegans
# post-intervention: 20% vegans
# hence the absurdly large point estimate

escalc_add_row( authoryear = "30-Day VeggieChallenge (ProVeg) 2017",
                substudy = NA,
                desired.direction = 1,
                effect.measure = "log-rr",
                interpretation = "Self-describing as vegan after vs. before challenge",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "RR",

                ai = 697, # pre-intervention vegans
                bi = 4727+7502+2573,  # pre-intervention non-vegans
                ci = 799+1155+530+697,  # post- vegans
                di = 1239+2473+4689+217+1658+2043 ) # post-non-vegans

# sanity check
( (799+1155+530+697) / (799+1155+530+697+1239+2473+4689+217+1658+2043) ) / ( 697 / (697+4727+7502+2573) )



##### #3828 10 Weeks to Vegan (Vegan Outreach) #####
# MM audited 2020-2-5

# "Food Intake" table
escalc_add_row( authoryear = "10 Weeks to Vegan (Vegan Outreach) 2018",
                substudy = NA,
                desired.direction = 1,
                effect.measure = "log-rr",
                interpretation = "Being vegan after vs. before intervention",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "RR",

                ai = 43, # pre-intervention vegans
                bi = 190+76,  # pre-intervention non-vegans
                ci = 77,  # post- vegans
                di = 140+92 ) # post-non-vegans

# sanity check
43/(190+76+43)
77/(140+92+77)


###### #3839 - Great Vegan Challenge (Animal Aid) #####
# MM audited 2020-2-5

# from Grassian
# measure bar heights on page 10 ("Dietary Trends Over Time")

total.px = 850  # total vertical height representing 100%
bl.vegans.px = 31  # height of vegan part of the "0 months" bar
fu.vegans.px = 240  # height of the vegan part of the "6 months" bar
N = 470 # pg 6

( bl.prop.vegan = bl.vegans.px/total.px )
( fu.prop.vegan = fu.vegans.px/total.px )

escalc_add_row( authoryear = "The Great Vegan Challenge (Animal Aid) 2016",
                substudy = NA,
                desired.direction = 1,
                effect.measure = "log-rr",
                interpretation = "Being vegan after vs. before intervention",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "RR",

                ai = bl.prop.vegan * N, # pre-intervention vegans
                bi = (1-bl.prop.vegan) * N,  # pre-intervention non-vegans
                ci = fu.prop.vegan * N,  # post- vegans
                di = (1-fu.prop.vegan) * N ) # post-non-vegans


##### #3840 Veganuary (Veganuary) 2014 #####
# MM audited 2020-2-5

# % staying vegan, % reducing consumption of each category
# both broken down by baseline diet

# not perfect comparison due to reporting limitations,
#  but am comparing the proportion who said they were vegan before challenge
#  to proportion saying they ate only vegan food during the challenge

# no pages in this document, but PDF has relevant stats circled

escalc_add_row( authoryear = "Veganuary (Veganuary) 2014",
                substudy = NA,
                desired.direction = 1,
                effect.measure = "log-rr",
                interpretation = "Being vegan after vs. during intervention",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "RR",

                ai = .149 * 3325, # pre-intervention vegans
                bi = (1-.149) * 3325,  # pre-intervention non-vegans
                ci = 330,  # post- vegans
                di = 711-330 ) # post-non-vegans


##### #3847, #3849 - Veganuary (Veganuary) 2016 #####
# MM audited 2020-2-5

# #3849 is a Fauna report, so more info available
# Table on page 3
escalc_add_row( authoryear = "Veganuary (Veganuary) 2016",
                substudy = NA,
                desired.direction = 1,
                effect.measure = "log-rr",
                interpretation = "Being vegan after vs. before intervention",
                use.rr.analysis = 1,
                use.grams.analysis = 0,
                use.veg.analysis = 0,
                measure = "RR",

                ai = .19 * 20597, # pre-intervention vegans
                bi = (1-.19) * 20597,  # pre-intervention non-vegans
                ci = .54 * 3369,  # post- vegans
                di = (1-.54) * 3369 ) # post-non-vegans

###### **#3826 Challenge 22+ (Animals Now) #####
# MM audited 2020-2-5

d = dplyr::add_row(.data = d,
                   authoryear = "Challenge 22+ (Animals Now) 2018",
                   substudy = NA,
                   desired.direction = 1,
                   effect.measure = "log-rr",
                   interpretation = "Being vegan after vs. before intervention",
                   use.rr.analysis = 1,
                   use.grams.analysis = 0,
                   use.veg.analysis = 0,
                   yi = 0.0739,
                   vi = 0.0001 )

##### #3853, #3854 - Veganuary (Veganuary) 2015 #####
# MM audited 2020-2-5
# 49% stayed vegan after six months

##### #3846 - Veganuary (Veganuary) 2017 #####
# MM audited 2020-2-5
# 67% intending to stay vegan

##### #3845 - Veganuary (Veganuary) 2018 #####
# MM audited 2020-2-5
# 62% intend to stay vegan

##### #3841 - Veganuary (Veganuary) 2019 #####
# MM audited 2020-2-5
# 47% intending to stay vegan

##### #3833, #3835 - Summer Vegan Pledge (Animal Aid) 2018 #####
# MM audited 2020-2-5
# 53% said they have remained vegan

##### #3834 Summer Vegan Pledge (Animal Aid) 2019 #####
# MM audited 2020-2-5
# 57% said they would remain vegan




# save intermediate dataset
setwd(data.dir)
write.csv(d, "data_prepped_step2.csv", row.names = FALSE)

} # end another "if (reenter.effect.sizes == TRUE)" statement

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ #
#                                     STEP 3 - MERGE IN QUALITATIVE DATA AND PREP ANALYSIS VARIABLES                                      #
# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ #

# MM audited 2020-2-5

# read it back in
setwd(data.dir)
d = read.csv("data_prepped_step2.csv")
expect_equal( nrow(d), 108 )  # includes SSWS

# how many unique studies?
length(unique(d$authoryear))
# how many point estimates?
nrow(d)


# qualitative data entered into Excel
setwd(data.dir)
# NOTE: this step breaks if cell values are hyphenated!
d2 = read_xlsx("Extracted qualitative data.xlsx", na = "NR")
# remove missing rows, used for human-readability
d2 = d2 %>% filter(!is.na(`First author last name`))


# for some reason, reads in years in an absurd format (e.g., "2018.0" as a string)
library(tidyverse)
d2$Year = str_remove(d2$Year, "[.]0")

# clean up carriage returns
d2$`Subject country` = str_replace_all(d2$`Subject country`, "[\n]" , "")


##### unique merger variable
d$unique = NA
d$unique[ is.na(d$substudy) ] = as.character(d$authoryear[ is.na(d$substudy) ])
d$unique[ !is.na(d$substudy) ] = paste( d$authoryear[ !is.na(d$substudy) ],
                                        d$substudy[ !is.na(d$substudy) ],
                                        sep = " ")
d$unique

# make pretty for eventual forest plot
d2$unique = NA
d2$unique[ is.na(d2$`Substudy #`) ] = paste( d2$`First author last name`[ is.na(d2$`Substudy #`) ],
                                             d2$Year[ is.na(d2$`Substudy #`) ],
                                             sep = " ")
d2$unique[ !is.na(d2$`Substudy #`) ] = paste( d2$`First author last name`[ !is.na(d2$`Substudy #`) ],
                                              d2$Year[ !is.na(d2$`Substudy #`) ],
                                              d2$`Substudy #`[ !is.na(d2$`Substudy #`) ],
                                              sep = " ")
d2$unique

# look for IDs from effect sizes that aren't in the qualitative data spreadsheet
#  (should be none)
expect_equal( d$unique[ !d$unique %in% d2$unique ], character(0) )

# studies in qualitative data for which we haven't entered effect sizes
# should occur only for hopeless studies (i.e., impossible to get stats)
known.hopeless = c( "Dowsett 2018",
                    #"Diaz 2019",  # removed from spreadsheet because we couldn't determine its eligibility
                    "de Lanauze 2019",
                    "Summer Vegan Pledge (Animal Aid) 2018",
                    "Summer Vegan Pledge (Animal Aid) 2019",
                    "Veganuary (Veganuary) 2017",
                    "Veganuary (Veganuary) 2018",
                    "Veganuary (Veganuary) 2019" )
expect_equal( sort(d2$unique[ !d2$unique %in% d$unique ]), sort(known.hopeless) )

# merge them
d = merge( d,
           d2,
           all.x = TRUE,
           by.x = "unique",
           by.y = "unique" )


############################### MERGE IN SUBJECTIVELY-RATED QUALITY DATA ###############################

# MM audited 2020-2-5

# read in subjective quality data
setwd(data.dir)
setwd("Dual review of quality")
d3 = read.csv("subjective_data_full_prepped.csv")

# check for studies lacking entries in subjective data
# should be only SSWS
d$authoryear[ !d$authoryear %in% d3$authoryear ]

# merge with main dataset
d = merge( d,
              d3[ , c("authoryear", "qual.exch", "qual.sdb", "qual.gen") ],
              all.x = TRUE,
              by = "authoryear" )


############################### MERGE IN INTERVENTION COMPONENTS DATA ###############################


setwd("~/Dropbox/Personal computer/Independent studies/2019/AWR (animal welfare review meat consumption)/Linked to OSF (AWR)/Data extraction/Dual review of intervention components")

d4 = read_xlsx("*component_coding_reconciled.xlsx")
d4 = d4 %>% filter( !is.na(authoryear) ) # remove spacer row

expect_equal( nrow(d4), 100 )

# rename columns 
d4 = rename(d4, x.mind.attr = "mind.attribution",
            x.soc.norm = "social.norms",
            x.id.victim = "id.victim",
            x.impl = "impl.suggest",
            x.pets = "pets")

# recode "Unclear" as NA
vars = c("x.mind.attr", "x.soc.norm", "x.id.victim", "x.impl", "x.pets")
d4 = d4 %>% replace_with_na_all( condition = ~.x == "Unclear" ) 
d4 = d4 %>% mutate_at( vars, as.numeric )
# sanity check (compare to before recoding "Unclear")
# table(d4$x.soc.norm, useNA = "ifany")

# check for studies lacking entries in subjective data
# should be only SSWS studies
d$authoryear[ !d$authoryear %in% d4$authoryear ]

# merge with main dataset
d = merge( d,
           d4[ , c("unique", vars) ],
           all.x = TRUE,
           by = "unique" )

# should still have same size
expect_equal( nrow(d), 108 )

# save intermediate dataset
setwd(data.dir)
write.csv(d, "data_prepped_step3.csv", row.names = FALSE)



# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ #
#                          STEP 4 - CONVERT EFFECT SIZES TO RR SCALE                                      #
# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ #

# read it back in
# check.names = FALSE retains spaces in variable names
setwd(data.dir)
d = read.csv("data_prepped_step3.csv", check.names = FALSE)

############################### CONVERT EFFECT SIZES - to RRs ###############################
# MM audited 2020-2-5

# types of effect measures 
table( d$effect.measure )

# synchronize directions so that positive is always good
d$yi[ d$desired.direction == 1 ] = abs(d$yi[ d$desired.direction == 1 ])
d$yi[ d$desired.direction == 0 ] = -abs(d$yi[ d$desired.direction == 0 ])

# init estimates on scale used for analysis
d$logRR = NA
d$varlogRR = NA

##### RRs #####
# no conversion needed, obviously
d$logRR[ d$effect.measure == "log-rr" ] = d$yi[ d$effect.measure == "log-rr" ]
d$varlogRR[ d$effect.measure == "log-rr" ] = d$vi[ d$effect.measure == "log-rr" ]

##### SMDs #####
RR.stats = d_to_logRR( smd = d$yi[ d$effect.measure == "smd" ],
                       smd.se = sqrt(d$vi[ d$effect.measure == "smd" ]) )
d$logRR[ d$effect.measure == "smd" ] = RR.stats$logRR
d$varlogRR[ d$effect.measure == "smd" ] = RR.stats$varlogRR


# sanity check for absurd values
round( sort( exp(d$logRR) ), 2 )
sort(sqrt(d$varlogRR))
round( sort(d$logRR/sqrt(d$varlogRR)), 2 )  # z-scores


##### Calculate Approximate CI limits for Forest Plot #####
z.crit = qnorm(.975)
d$RR.lo = exp( d$logRR - z.crit * sqrt(d$varlogRR) )
d$RR.hi = exp( d$logRR + z.crit * sqrt(d$varlogRR) )


##### Affirmative vs. Nonaffirmative #####
d$pval = 2 * ( 1 - pnorm( abs(d$logRR) / sqrt(d$varlogRR) ) )
d$affirm = d$pval < 0.05 & d$logRR > 0
table(d$affirm)

# sanity check
# fine for RMDs to be NA since those are for the grams analysis
table( is.na(d$logRR), d$effect.measure)

# don't use these for the non-main-RR analyses
d$logRR[ d$use.rr.analysis == 0 ] = NA  
d$varlogRR[ d$use.rr.analysis == 0 ] = NA
d$RR.lo[ d$use.rr.analysis == 0 ] = NA
d$RR.hi[ d$use.rr.analysis == 0 ] = NA


# save intermediate dataset
setwd(data.dir)
write.csv(d, "data_prepped_step4.csv", row.names = FALSE)


# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ #
#                                  STEP 5 - RECODE VARIABLES                                     #
# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ #

# read it back in
# check.names = FALSE retains spaces in variable names
setwd(data.dir)
d = read.csv("data_prepped_step4.csv", check.names = FALSE)

############################### MAKE NEW VARIABLES AND RENAME THE EXISTING ONES ###############################

# MM audited 2020-2-5

d = d %>%
  rename(
    # meta-information about study and article
    year = Year,
    title = Title,
    ref = `Ref #`,
    journal = `Journal/conference (if peer-reviewed)`,
    other.source = `Other source (if not peer-reviewed)`,
    exclude.main = `Excluded challenge`,
    borderline = `Borderline inclusion`,
    mm.fave = `Among MM's favorites methodologically, exclusive of small sample size`,
    stats.source = `Stats source (public data, data from author, paper, hopeless)`,
    
    # study design characteristics
    perc.male = `Percent male`,
    age = `Mean or median age`,
    students = `Students ("No", "General undergraduates", "Social sciences undergraduates", "Mixed")`,
    country = `Subject country`,
    design = Design,
    n.paper = `N (total analyzed sample size in paper, combining all substudies included here)`,
    
    # intervention characteristics
    x.has.text = `Intervention has text`,
    x.suffer = `Intervention has specific description or images of animal suffering`,
    x.has.visuals = `Intervention has visuals`,
    x.pure.animals = `Intervention is purely animal welfare`,
    x.suffer = `Intervention has specific description or images of animal suffering`,
    x.rec = `Intervention recommendation (reduce, go vegan, go vegetarian, no request, other)`,
    x.tailored = `Intervention personally tailored`,
    x.min.exposed = `Total time exposed to intervention (minutes)`,
    # subjective ones that were dual-coded:
    x.mind.attr = x.mind.attr,
    x.soc.norm = x.soc.norm,
    x.id.victim = x.id.victim,
    x.impl = x.impl,
    x.pets = x.pets,
    
    # outcome characteristics
    y.cat = `Outcome category (purchase or consumption)`,
    y.lag.days = `Outcome time lag from intervention (days)`,
    y.other.eligible.foods = `Other food outcomes available?`,
    
    # quality variables
    qual.y.prox = `Outcome proximity (intended, self-reported, actual)`,
    qual.missing = `Missing data (%)`,
    qual.prereg = Preregistered,
    qual.public.data = `Public data`,
    qual.public.code = `Public code`,
    
    # prose
    prose.x = `Intervention (X)`,
    prose.control = `Control condition`,
    prose.outcome = `Outcome (Y)`,
    prose.population = `Population`,
    prose.y.other.eligible = `Additional eligible food outcomes`,
    prose.notes = `Notes`
  )

# get rid of columns with capital letters (not used in analysis)
has.caps = tolower(names(d)) != names(d)
d = d[ , has.caps == FALSE | names(d) %in% c("logRR", "varlogRR", "RR.lo", "RR.hi") ]


############################### RECODE VARS AND MAKE NEW ONES ###############################

# MM audited 2020-2-5

##### Cast As Numeric #####
make.numeric = c("perc.male",
                 "x.min.exposed",
                 "qual.missing")
d = d %>%
  mutate_at( make.numeric, as.numeric )


##### Simple Recodings #####
d$published = !is.na(d$journal)
d$y.lag.wks = d$y.lag.days/7
d$y.long.lag = d$y.lag.days >= 7
d$randomized = grepl(" RCT", d$design)  # the leading space prevents "NRCT" from counting as randomized
d$reproducible = (d$qual.prereg == "Yes") & (d$qual.public.data == "Yes")
d$x.long = d$x.min.exposed >= 5
# recode percent male as a 10-percentage point increase
d$perc.male.10 = d$perc.male/10
# recode age as a 5-year increase
d$age.5 = d$age/5

##### Recode Categorical Variables to Force Levels #####
library(car)
d$x.rec = recode_factor( d$x.rec,
                           "No request" = "a.No request",
                           "Reduce" = "b.Reduce",
                           "Go vegetarian" = "c.Go vegetarian",
                           "Go vegan" = "d.Go vegan",
                           "Mixed" = "e.Mixed")

d$qual.y.prox = dplyr::recode( d$qual.y.prox,
                               "Actual" = "a.Actual",
                               "Self-reported" = "b.Self-reported",
                               "Intended" = "c.Intended" )

# collapse categories
d$qual.y.prox2 = recode_factor( d$qual.y.prox,
                                "a.Actual" = "b.Actual or self-reported",
                                "b.Self-reported" = "b.Actual or self-reported",
                                "c.Intended" = "a.Intended")



# collapse categories
# any request vs. no request
d$x.makes.request = dplyr::recode( d$x.rec,
                                   "a.No request" = 0,
                                   "b.Reduce" = 1,
                                   "c.Go vegetarian" = 1,
                                   "d.Go vegan" = 1,
                                   "e.Mixed" = 1)

d$design = recode_factor( d$design,
                           "Between-subjects RCT" = "a.Between-subjects RCT",
                           "Within-subject RCT" = "b.Within-subject RCT",
                          "Cluster RCT" = "c.Cluster RCT",
                           "Between-subjects NRCT" = "d.Between-subjects NRCT",
                          "Within-subject UCT" = "e.Within-subject UCT")

# rename in a risk-of-bias way
d$qual.sdb = dplyr::recode( d$qual.sdb,
                                   "strong" = "a.Low",
                                   "medium" = "b.Medium",
                                   "weak" = "c.High",
                                   "unclear" = "d.Unclear" )
d$qual.exch = dplyr::recode( d$qual.exch,
                             "strong" = "a.Low",
                             "medium" = "b.Medium",
                             "weak" = "c.High",
                             "unclear" = "d.Unclear" )
d$qual.gen = dplyr::recode( d$qual.gen,
                            "strong" = "a.Low",
                            "medium" = "b.Medium",
                            "weak" = "c.High",
                            "unclear" = "d.Unclear" )

# recode missing data
d[ d == "NR" ] = NA
d[ d == "na" ] = NA


############################### SIMPLIFIED "TABLE 1" DATASET ###############################

# MM audited 2020-2-5

d.small = d %>% dplyr::select( unique, 
                               authoryear,
                               substudy,
                               title,
                               journal,
                               other.source,
                               borderline,
                               exclude.main,
                               prose.population,
                               country,
                               n.paper,
                               perc.male,
                               prose.x,
                               prose.control,
                               x.tailored,
                               x.has.text,
                               x.has.visuals,
                               x.suffer,
                               x.pure.animals,
                               x.mind.attr,
                               x.soc.norm,
                               x.id.victim,
                               x.impl,
                               x.pets, 
                               x.min.exposed,
                               x.rec,
                               prose.outcome,
                               y.cat,
                               y.other.eligible.foods,
                               prose.y.other.eligible,
                               
                               design,
                               randomized,
                               qual.y.prox,
                               qual.missing,
                               qual.prereg,
                               qual.public.data,
                               qual.public.code,
                               qual.exch,
                               qual.gen,
                               qual.sdb,
                               
                               effect.measure,
                               desired.direction,
                               yi,
                               vi,
                               logRR,
                               varlogRR,
                               RR.lo,
                               RR.hi ) 

############################### WRITE PREPPED DATA ###############################

setwd(data.dir)
# analysis dataset
write.csv(d, "prepped_data.csv", row.names = FALSE)

# same as above, but leaner and meaner choice of variables
write.csv(d.small, "prepped_data_lean.csv", row.names = FALSE)
